🔀 chore(dashboard-api): update tsconfig.lib.json to include renameFolderQrCodes action file
🔀 chore(dashboard): move renameFolderQrCodes action file to actions folder

🔀 chore(dashboard): add renameFolderQrCodes action file

🔀 chore(pass-api): remove unused exports from common file

🔀 chore(pass-api): update import path for getPassUser in downloadPass file

🔀 chore(pass-api): update import path for getPassUser in revealPass file

🔀 chore(pass-api): update import path for getPassUser in transferPassQrCode file

🔀 chore(pass-api): update import path for getPassUser in transferPassQrCodeBatch file

🔀 chore(pass-common): add eslint configuration file

📦 chore(pass-common): add pass-common library to the project
📦 chore(pass-common): add swcrc configuration file for pass-common library
📦 chore(pass-common): add README.md file for pass-common library
📦 chore(pass-common): add jest.config.ts file for pass-common library
📦 chore(pass-common): add package.json file for pass-common library
📦 chore(pass-common): add project.json file for pass-common library
📦 chore(pass-common): add index.ts file for pass-common library
📦 chore(pass-common): add tsconfig.json file for pass-common library
📦 chore(pass-common): add tsconfig.lib.json file for pass-common library
📦 chore(pass-common): add tsconfig.spec.json file for pass-common library
📦 chore(pass-common): add pass-common library to tsconfig.base.json
📦 chore(package.json): add @sentry/utils dependency
