+++ libs/features/back-office/dashboard-api/src/index.ts
@@ -1,4 +1,3 @@
 export * from './lib/createEventPassNftContract';
 export * from './lib/createEventPassNfts';
 export * from './lib/getEventsFromOrganizerId';
-export * from './lib/renameFolderQrCodes';
