+++ libs/features/back-office/dashboard-api/tsconfig.lib.json
@@ -19,5 +19,11 @@
     "src/**/*.spec.jsx",
     "src/**/*.test.jsx"
   ],
-  "include": ["src/**/*.js", "src/**/*.jsx", "src/**/*.ts", "src/**/*.tsx"]
+  "include": [
+    "src/**/*.js",
+    "src/**/*.jsx",
+    "src/**/*.ts",
+    "src/**/*.tsx",
+    "../dashboard/src/lib/actions/renameFolderQrCodes.ts"
+  ]
 }
