+++ libs/features/pass-api/src/index.ts
@@ -1,8 +1,3 @@
 export * from './lib/transferPassQrCodeBatch';
 export { getEventPassNftByTokenReferenceUser } from './lib/getEventPassNftByTokenReferenceUser';
 export { getEventPassNftByTokenReferenceAnonymous } from './lib/getEventPassNftByTokenReferenceAnonymous';
-export {
-  getPassOrganizer,
-  getPassUser,
-  getEventPassOrganizerFolderPath,
-} from './lib/common';
