+++ libs/features/pass-common/README.md
@@ -1,0 +1,11 @@
+# features-pass-common
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build features-pass-common` to build the library.
+
+## Running unit tests
+
+Run `nx test features-pass-common` to execute the unit tests via [Jest](https://jestjs.io).
