+++ libs/features/pass-common/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@features/pass-common",
+  "version": "0.0.1"
+}
