+++ tsconfig.base.json
@@ -52,6 +52,7 @@
       ],
       "@features/pass": ["libs/features/pass/src/index.ts"],
       "@features/pass-api": ["libs/features/pass-api/src/index.ts"],
+      "@features/pass-common": ["libs/features/pass-common/src/index.ts"],
       "@features/pass-types": ["libs/features/pass-types/src/index.ts"],
       "@features/pass/server": ["libs/features/pass/src/server.ts"],
       "@features/settings": ["libs/features/settings/src/index.ts"],
